pyproject.toml
[tool.poetry]
name = "chungang-univ-notice-bot"
version = "0.1.0"
description = "Chung Ang University Notice Bot"
authors = ["Junu Park <junupark1@cau.ac.kr>"]
readme = "README.md"
packages = [
    { include = "src" }
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.9.3"
requests = "^2.28.1"
python-telegram-bot = "^13.7"
google-cloud-secret-manager = "^2.12.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.0.0"
flake8 = "^6.0.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry-core.backend"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"